backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
alembic
psycopg[binary]
redis
requests
PyJWT
